+++ src/REslava.ResultFlow/Generators/ResultFlow/CodeGeneration/ResultFlowMermaidRenderer.cs
@@ -206,7 +206,7 @@ namespace REslava.ResultFlow.Generators.ResultFlow.CodeGeneration
                     {
                         var fId = $"F{failureCounter++}";
                         var gatekeeperLabel = node.PredicateText != null
-                            ? "<span title='" + node.PredicateText.Replace("\"", "&quot;").Replace("'", "&#39;") + "'>" + label + "</span>"
+                            ? "<span title='" + node.PredicateText.Replace("\"", "\u201c").Replace("'", "\u2019") + "'>" + label + "</span>"
                             : label;
                         lines.Add($"{indent}{nodeId}[\"{gatekeeperLabel}\"]:::gatekeeper");
                         if (hasNext) lines.Add($"{indent}{nodeId} -->|pass| {resolvedNextId}");